refactored the jena based rules service business logic into a generic rules based reasoning service class and updated application configuration to instantiate a bean using the relevant ruleset
