+++ src/main/java/darkdata/service/RuleBasedCompatibilityService.java
@@ -1,31 +1,18 @@
 package darkdata.service;
 
-import darkdata.datasource.DarkDataDatasource;
 import darkdata.model.kb.PhysicalFeature;
 import darkdata.model.kb.candidate.CandidateWorkflow;
 import darkdata.model.kb.compatibility.CompatibilityAssertion;
 import darkdata.model.kb.compatibility.CompatibilityValue;
 import darkdata.model.kb.g4.G4Service;
 import darkdata.model.ontology.DarkData;
-import org.apache.jena.base.Sys;
 import org.apache.jena.ontology.OntModel;
 import org.apache.jena.ontology.OntProperty;
 import org.apache.jena.rdf.model.InfModel;
-import org.apache.jena.rdf.model.ModelFactory;
 import org.apache.jena.rdf.model.RDFNode;
-import org.apache.jena.reasoner.rulesys.GenericRuleReasoner;
-import org.apache.jena.reasoner.rulesys.Rule;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
-import org.springframework.util.Assert;
 
-import javax.annotation.PostConstruct;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -38,35 +25,13 @@ import java.util.stream.Stream;
 @Service
 public class RuleBasedCompatibilityService implements CandidateWorkflowCompatibilityService {
 
-    @Value("classpath:rules/some.rules")
-    private Resource ruleset;
-
     @Autowired
-    private DarkDataDatasource datasource;
-
-    private GenericRuleReasoner reasoner;
-
-    @PostConstruct
-    public void init() throws IOException {
-        try (InputStream is = this.ruleset.getInputStream()) {
-            BufferedReader reader = new BufferedReader(new InputStreamReader(is));
-            this.reasoner = new GenericRuleReasoner(Rule.parseRules(Rule.rulesParserFromReader(reader)));
-            reasoner.bindSchema(datasource.getSchema());
-            reasoner.setDerivationLogging(true);
-        }
-    }
-
-    public GenericRuleReasoner getReasoner() {
-        return reasoner;
-    }
+    private RuleBasedReasoningService basicRulesReasoningService;
 
     @Override
     public List<CompatibilityAssertion> computeCompatibilities(CandidateWorkflow candidate) {
         OntModel m = candidate.getIndividual().getOntModel();
-        InfModel ruleInf = ModelFactory.createInfModel(getReasoner(), m);
-        ruleInf.add(datasource.getSchema());
-
-        ruleInf.prepare();
+        InfModel ruleInf = basicRulesReasoningService.reason(m);
 
         // TODO run these 5 computations in parallel
         computeStrongCompatibilityAssertions(ruleInf, candidate);
@@ -76,7 +41,6 @@ public class RuleBasedCompatibilityService implements CandidateWorkflowCompatibi
         computeNegativeCompatibilityAssertions(ruleInf, candidate);
 
         return candidate.getCompatibilityAssertions();
-
     }
 
     public void computeStrongCompatibilityAssertions(InfModel ruleInf, CandidateWorkflow candidate) {
